MovePage: clear title cache before parsing new redirect.
Clear all caches to make sure no stale information is used
when parsing the newly created redirect. Without this, moves would fail
under certain conditions when Lua core runs on the new page.
It is not entirely clear why this is needed, but it seems to fix the
issue when I tried it on mwdebug.

Bug: T279832
Change-Id: I222fe577c907aa9e6ac9accc597b3a1e45c9d7cc