Legacy feature should not load thumbnail style rules (only layout)
Previously thumbnail styles were split between the legacy and content
modules. The styles in legacy were concerned with layout only for example
floating and text alignment.

When migrating styles from legacy to a new feature called
`content.thumbnails` for backwards compatible we imported the entire
feature. On hindsight this was a mistake.
This had the impact of loading thumbnail styles on skins that did
not want those styles, with Timeless being the most known impact.

It is proposed we reconsider the patch in c504f19f1fa and instead
give proper backwards compatibility to skins.

Bug: T280292
Change-Id: I4464eb9195f63f27ea56d9b7baff02b45f3c3a13