selenium: Record videos using Xvfb+ffmpeg
Unil 2020-06-15 we were using Xvfb+ffmpeg to record videos.

In eed975d we have switched to wdio-video-reporter npm package.
It has never worked very well. It does not really record videos.
It creates screenshots and creates videos from them.
It does not record modal popups. Sometimes, it doesn't create a video
for a test at all.

I think it's time to switch back to Xvfb+ffmpeg. It has worked well. The
reason we moved to wdio-video-reporter was that it worked on all
platforms. That's not really important any more, since Xvfb+ffmpeg works
in a Fresh container.

Bug: T253113
Change-Id: I221775208576b6496f682ceefd6e6334db488bd4