preferences: Move complex creation of infos for usergroups into closure
The HTMLInfoField allows the default part to be a closure.
This delays the message parse until the messages are needed on a gui.
This avoids the message parse on save or api calls and hopefully makes
that faster.

The message prefs-memberingroups is now parsed as all the other label
messages. This is another fix as done for T142882 / 116063f

Bug: T58633
Change-Id: I860501fe3df1ee330460f04ffe8493dd16421c8d