Convert NamespaceDefTest to a unit test
Most other TypeDefTestCase subclasses are already
unit tests (TypeDefTestCase does not extend
MediaWikiIntegrationTestCase, but rather
\PHPUnit\Framework\TestCase directly) but
NamespaceDefTest was treated as an integration test
based on its location so that it could use MediaWikiServices
for the NamespaceInfo. Instead, use the NamespaceInfo
from DummyServicesTrait. Requires passing an explicit
mock HookContainer because otherwise DummyServicesTrait
would call ->createHookContainer() which is only available
for MediaWikiTestCaseTrait

Change-Id: I279ae84702ba329ea54efb39e996e9705710c706