ApiQueryUserContribs: Clarify the labels for ucstart and ucend
The documentation labels for `ucstart` and `ucend` are misleading because contributions are shown from later to earlier. This PR adds clarification to the labels, to inform the API user that they need to provide the timestamp before the revisions for `ucstart` and timestamp after the revisions in `ucend`.

Does not include i18n.

Bug: T262834
Change-Id: I85af0772a412345aa3b5fe51911481be491584b9