Add getQueryInfo to RevisionFactory
RevisionFactory is typically used when selecting from the revisions or
archive table, and then constructing RevisionRecords from the resulting
rows. Since RevisionFactory expects certain fields to be present in the
row, it makes sense to have the getQueryInfo methods that define these
lists exposed via the same interface as well.

Bug: T282844
Change-Id: I16908a4fcbdf72d4b765614d628bee5c6ce278eb