Add a composer command for the custom PHPUnit entry point
This is meant to ease migration from the custom entry point to the
standard vendor/bin/phpunit in a fashion that is essentially invisible
to any consumer of this command. So after this change is merged we can
change everything that uses tests/phpunit/phpunit.php to use this
composer shortcut, and then eventually make the shortcut point to
vendor/bin/phpunit with the right configuration.

Bug: T90875
Change-Id: I2388610cda4fcbe87a4187c04ff624db5fa4d17c