mediawiki.page.ready: Introduce `wikipage.indicators` hook
This new hook allows extensions and gadgets to enhance indicators on
page load and on live preview. This is useful for the Kartographer
extension.

The hook gets fired only if there is at least one indicator present.

The new hook gets fired before `wikipage.content` fires, to have the
same order like in the DOM.

Bug: T280363
Change-Id: I154ebbf049fbce1e872548651953adb1419b6cba