Fix array order for array_replace_recursive merge strategy
Prior to this change, if a configuration array is defined
in extension.json, array elements cannot be overridden by
setting values in the corresponding configuration variable,
as intended by the array_replace_recursive merge strategy.
This is because the order of the arrays in the
array_replace_recursive() function call need to be
reversed.

Bug: T283464
Change-Id: I55561232a994f745c5f3cd8394674b18c0529d13
(cherry picked from commit 4d413376b96b2a377f7a5206f63742743bd3f809)