Dabatase: Assert that join conditions are arrays
Join conditions are in the form of [ <table> => [ 'JOIN', <cond> ] ].
If they are given incorrectly as strings, e.g. in the form
[ <table> => 'JOIN', <cond> ], there won't be any error (courtesy
of PHP handling list($a,$b)='x' by setting $a and $b to null) and
the resulting SQL will probably even be valid, just completely
different from what's intended. Be more strict about verifying
the format.

Change-Id: Id09aff6193829972cdca3262dc8bb878512c6cc3