Update wvui to 0.2.0
Use the wvui-search bundle for now, which excludes the new components
that use the composition API plugin. Once the security readiness review
for the composition API plugin (T281527) is completed, we can use the
full bundle for the 'wvui' module, and create a new 'wvui-search' module
that uses the smaller search bundle.

Bug: T280982
Bug: T277845
Bug: T280840
Bug: T279015
Bug: T280827
Bug: T277534
Bug: T280934
Bug: T281362
Bug: T282937
Bug: T281186
Bug: T281797
Change-Id: Ia5e14fb9d0073a5126a0918f7a94213c671e773a