Revert "LocalFile: don't use dynamic property assignment when loading from the DB"
This reverts commit 29cd484c864a8cf84a30a6ec4850b8caa95caf1e.

Reason for revert: subclasses of LocalFile broke, cause they were relying on dynamic property assignment. For example OldLocalFile was loading oi_deleted this way. We could override loadFromDB, but since LocalFile is @stable to override, this becomes a breaking change. I'll leave it to Tim to re-revert in a way he thinks it's best.

Change-Id: I81e06c13128d3330972a39fb2628a1dd5f8c4658