Remove workarounds for IE8 from TOC style
The IE8 don't support the CSS pseudo-class `:checked` and therefor
matches CSS selectors with the prefix `:not( :checked ) >` never
on IE8 and always on newer browsers with support of `:checked`.
This was used to hide styles for IE8.

IE8 was the last browser that doesn't support `:checked` and the basic
support of IE8 was removed in MediaWiki 1.36.

Bug: T248061
Bug: T258766
Change-Id: Ie68e9dae441fe1dab57026c451d198c4ba332702