mediawiki.viewport: Replace ambiguous "[1.37]" mark with words
This novel syntax has not been used before and seems ambigious in
meaning in the context of the browser console.

Also remove the duplicate use of the module name, which we don't
mention again in deprecation messages. The system does this already.
E.g. "deprecated X used" + message with version since and/or what to
use instead.

Bug: T284179
Change-Id: I3158134daab21fc4268448c9ea0afcadbae9b27f