filerepo: Fix parsing metadata in ForeignAPIFile
ForeignAPIFile::parseMetadata is a recursive function trying to parse the
metadata tree. We want to emit _error as value if the metadata returned
by API is not an array but at the same time, we broke when the function
reaches the leaves and tries to parse them making parseMetadata return _error
for all leaves instead.

This broke beta cluster. I tested this change on beta cluster and this
patch fixes it.

Bug: T275268
Change-Id: I2befa92a8f9619ba6eb990d461f0a6f0fa7a4677