Let ApiSandbox display non-JSON responses like Xdebug errors
When developing locally, an API response might not be valid
JSON but start with e.g. an Xdebug stacktrace, formatted as
an HTML table. As of now this makes the sandbox fail with a
confusing exception. Let's just display the response instead.

While this approach feels like it might allow some kind of
injection, I don't think this is an issue.
1. There are already other places in this code that display
   HTML from the response.
2. The same HTML would be displayed anyway when the API is
   called without the sandbox.

Change-Id: Ic13e314f20f5b4cd6500f8e34168c25035358821