filerepo: Allow shared caching of ForeignAPIRepo responses across wikis
Currently, ForeignAPIRepo caches image metadata and siteinfo in
memcached, with the keys being local to the client wiki. This can result
in a low hit ratio if the same ForeignAPIRepo is shared between multiple
wikis. Since the cache keys for image metadata and site info already
vary on the foreign URL being fetched, it should be possible to make
these cache keys global and thereby allow reusing them across wikis that
share the same file repo configuration.

This is similar to what ForeignDBRepo/ForeignDBFile do already, although
note that for those classes the getSharedCacheKey() calls are not in
their own source, but inherited from LocalRepo/LocalFile.

Bug: T285271
Change-Id: Iaa20a27893e9748cb42fa3a3beb7cd4669b461cb