Allow a retry of flaky selenium test
Flaky selenium tests are a problem (https://w.wiki/3XMD).

Ideally we'd have better shared practices around writing tests that
reduce flakiness, but there will probably always be some element of
instability due to the nature of these tests.

We should fix flaky tests where there are clear fixes to do so, but it
would also be nice to not have to re-run a 20 minute build when a 5
second test, when retried, passes.

https://webdriver.io/docs/retry indicates that you can specify a number
of times to retry a spec file, so that's what is added here.

Change-Id: I29a4943e7ddaf755b242756543d9402177a0f6e5