SpecialChangeEmail: Requiring email confirmation is not an error
The message here reads "A confirmation email has been sent...".
It was styled as an error, but it's not an error message.
It's not quite a success message either (the user is required
to click a link in the email to continue), so let's style it
as a warning, I guess.

Change-Id: I85cb61e7da61d06b5fecd4e222f614820dcde688