SqlBagOStuff: Fix incorrect numServerShards for default CACHE_DB
As part of working on I, I noticed that, despite there being only
one shard on my localhost (getServerShardIndexes: ["local"]), only one
table index ([0]), and no data in them at all, purgeParserCache.php was
still, incorrectly, reporting the (only) progress callback from SqlBag
as being at 50%, with the final 100% only being printed due to the
CLI script forcing it so at the end - if not already done by SqlBag.

The following diff helped me understand this:

```
@@ -813,6 +814,10 @@ class SqlBagOStuff extends MediumSpecificBagOStuff {
+					echo json_encode( [
+						"tablesDoneRatio" => $tablesDoneRatio,
+						"numServerShards" => $this->numServerShards,
+					], JSON_PRETTY_PRINT );
 					call_user_func( $progressCallback, $overallRatio * 100 );
```

This should have printed 1 and 1, but printed 1 and 2. Thus the reason
it was only able to go upto 50% was because it wrongly thought there
was a second shard.

Change-Id: I9500a40d5c2a18750f043e33d825d748b79ae202