UserEditCountUpdate: only store a single UserIdentity object
No longer need to save more than one, which was previously
needed to be able to call User::setEditCountInternal(), but
that method has been removed.

It turns out we can't just rely on the user id alone, because
UserEditTracker requires a UserIdentity object and the name
of the user matters.

Follow-up:  Ie3fb4887d9f2d96b32598865030351bf3bf20ce5
Change-Id: I4ba1300837cf2cf624c563d4445a2a7dc6b88d86