DefaultSettings: Improve $wgUpdateCompatibleMetadata docs
The old description had some grammatical issues, and did not explain
what it influences or when. Namely, that it considered during manual
purges and overall rebuilds/refreshes. (E.g. it does not cause
update something directly on page views, nor does it kick-off a job
that will upgrade all existing files or something like that.)

Change-Id: Ia2d3f17a9f7c05f816004c619eb9ff1ec45b019b