Hard-deprecate some MWNamespace methods
MWNamespace has been deprecated in favor of the NamespaceInfo
service in 1.34 and should no longer be used.

All methods that are hard-deprecated in this patch are unused.

Also adding @deprecated to all methods so it's more clear that
they are deprecated. The @deprecated at the beginning of the
class can easily be overlooked.

Change-Id: Ic5726297807d7959453407267e852f7be96a3e05