objectcache: Fix SqlBagOStuff progress printing to account for server tags
Follows-up 0545fdc3af207, which added `--tag` to purgeParserCache.php
to purge only one of the server shards (by its tag).

However, the percentage printing was still based on the total count
instead of on the count of the local $shardIndexes array, thus the
progress reported was artifically slowed down to a low percentage (
e.g. 33.3% if there are three shards), and then it stops, with 100.0%
added to the end by the CLI script:

```
... 0.0%
[…]
... 33.3%
... 100.0%
```

Fix this to only count the servers being iterated.

Bug: T282761
Change-Id: Id87868ceb9b7980442d57551e728d9f0d2784e01