Add counter-style for the Meetei (mni) and Santali (sat) languages
The application of the style is done in the file
mediawiki.skinning/i18n-ordered-lists.less,
as it is with other languages. The values "meetei" and "santali"
for list-style-type are not built into standard CSS, and have
to be created using @counter-style. I put it in a separate
CSS file because otherwise LESS doesn't seem to be able
to process it.

Bug: T263998
Change-Id: Ic83030c36199dd8ec945d826f4cd42993d18fdb3