objectcache: normalize $exptime to a TTL in APCUBagOStuff/WinCacheBagOStuff
The apcu_* methods do not accept unix timestamps as expiry values.
Previously, if a unix timestamp, one hour in the future, was passed to
APCUBagOStuff::set(), then the apcu value was set to expire a number of
seconds in the future equal to that unix timestamp (e.g. many years).

Instead, convert absolute expirations to relative expirations (TTLs).
This logic uses the same function as RedisBagOStuff.

Bug: T286260
Change-Id: I6df484eaaed81fa4741a7de1e633e682362ffb85