Simplifications to PasswordResetTest
Remove unneeded clearing of hooks:
we create the instance of PasswordReset
with an injected hook container with no hooks.
(User::mailPasswordInternal is also run from
TemporaryPasswordPrimaryAuthenticationProvider
but that code path is never executed).

Use a StaticUserOptionsLookup instead of a mock.

Use mocks instead of retrieving from MediaWikiServices
were possible.

Reduce duplication in ::makeUsers() by adding a helper
function to create a registered user with a valid email.

Change-Id: I9d771322ee5cdc78fb9378964f367549938cca04